Change internal type to int64

+++ common/datavalues/src/data_value.rs
@@ -338,7 +338,7 @@ impl From<&DataType> for DataValue {
             DataType::List(f) => DataValue::List(None, f.data_type().clone()),
             DataType::Struct(_) => DataValue::Struct(vec![]),
             DataType::String => DataValue::String(None),
-            DataType::Interval(_) => DataValue::Float64(None),
+            DataType::Interval(_) => DataValue::Int64(None),
         }
     }
 }
