+++ query/src/servers/clickhouse/writers/query_writer.rs
@@ -310,9 +310,9 @@ pub fn to_clickhouse_block(block: DataBlock) -> Result<Block> {
                 DataType::Interval(unit) => {
                     let seconds = IntervalUnit::avg_seconds(unit.clone()) as f64;
                     let vs: Vec<f64> = column
-                        .f64()?
+                        .i64()?
                         .into_no_null_iter()
-                        .map(|c| *c / seconds)
+                        .map(|c| (*c as f64) / seconds)
                         .collect();
                     result.column(name, vs)
                 }
